Fix list of nodes on /etc/ha.d/haresources
Was setting the current node always as the default node, but this is not
the case in most situations.
Had to change the way the cookbook heartbeat is used to properly pass
a default node and to accept a list of nodes instead of always getting
them from chef server.